Greated new class Calculator
