+++ Calculator.java
@@ -1,67 +1,27 @@
-import java.util.Scanner;
-
 public class Calculator {
+    Main main = new Main();
 
-    public static void main(String[] args) {
-        System.out.println("Hello, i'm Integer Calculator");
-        Calculator calculator = new Calculator();
-        calculator.calculatorInteger();
-    }
-
-    Scanner in = new Scanner(System.in);
-
-    public void calculatorInteger(){
-        int triggerRepeat = 0;
-        while(triggerRepeat == 0){
-            System.out.println("Choose your operation: +, -, *, /");
-            System.out.println("For exit - input any different char");
-            String input = in.next();
-
-            switch(input){
-                case "+" : System.out.println("Result " + plus());
-                    break;
-                case "-" : System.out.println("Result " + minus());
-                    break;
-                case "*" : System.out.println("Result " + multiply());
-                    break;
-                case "/" : System.out.println("Result " + divide());
-                    break;
-                default: triggerRepeat = 1;
-            }
-        }
-    }
-
-    public int[] input(){
-        Scanner in = new Scanner(System.in);
-        int[] mas = new int[2];
-        System.out.println("Input first number");
-        mas[0] = in.nextInt();
-        System.out.println("Input second number");
-        mas[1] = in.nextInt();
-        return mas;
-    }
-
-    public int plus(){
+    public int plus() {
         System.out.println("Operation + ");
-        int[] inputs = input();
+        int[] inputs = main.input();
         return inputs[0] + inputs[1];
     }
 
-    public int minus(){
+    public int minus() {
         System.out.println("Operation - ");
-        int[] inputs = input();
+        int[] inputs = main.input();
         return inputs[0] - inputs[1];
     }
 
-    public int multiply(){
+    public int multiply() {
         System.out.println("Operation * ");
-        int[] inputs = input();
+        int[] inputs = main.input();
         return inputs[0] * inputs[1];
     }
 
-    public int divide(){
+    public int divide() {
         System.out.println("Operation / ");
-        int[] inputs = input();
+        int[] inputs = main.input();
         return inputs[0] / inputs[1];
     }
 }
